requirements.txt
psycopg2-binary>=2.9.3
pika>=1.2.0
pytest>=7.0.0
flask>=2.0.0
flask-cors>=3.0.10
python-dotenv>=0.20.0 